Style: 기본 셋팅 및 테스트완료git add . #7
